Creation of a QuittableApp

+++ packages/pymodaq/src/pymodaq/extensions/custom_ext.py
@@ -112,11 +112,12 @@ class CustomExt(CustomApp):
         return self._modules_manager
 
     def exit_runner_thread(self, duration : int = 5000):
-        self.runner_thread.quit()
-        terminated = self.runner_thread.wait(duration)
-        if not terminated:
-            self.runner_thread.terminate()
-            self.runner_thread.wait()
+        if self.runner_thread and self.runner_thread.isRunning():
+            self.runner_thread.quit()
+            terminated = self.runner_thread.wait(duration)
+            if not terminated:
+                self.runner_thread.terminate()
+                self.runner_thread.wait()
 
     def create_dashboard_toolbar(self,
                                  add_dashboard: bool = True,
